+++ InstaMarry/app/src/main/res/layout/fragment_personal_tab_layout.xml
@@ -3,15 +3,12 @@
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-
-    <GridView
-        android:id="@+id/gridView"
+    <ListView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:columnWidth="90dp"
-        android:numColumns="auto_fit"
-        android:verticalSpacing="10dp"
-        android:horizontalSpacing="10dp"
-        android:stretchMode="columnWidth"
-        android:gravity="center" />
+        android:layout_height="wrap_content"
+        android:id="@+id/exploreListView"
+        android:layout_gravity="center_horizontal"
+        android:divider="#b5b5b5"
+        android:dividerHeight="1dp"
+        android:listSelector="@drawable/list_selector"/>
 </LinearLayout>
